mental-health-nlp/requirements.txt
torch>=2.0.0
transformers>=4.30.0
opacus>=1.4.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
umap-learn>=0.5.3
nltk>=3.8.0
wordcloud>=1.9.0
